CTCP-4876: PR comments 3
